Add corpus from pact-examples, then add macros to create tests/benchmarks for each corpus file, and add to appropriate test/bench files
