+++ Cargo.toml
@@ -3,4 +3,5 @@ resolver = "2"
 members = [
     "lib",
     "bin",
+    "macros",
 ]
