+++ corpus/pact-examples/commercial-paper/auth.pact
@@ -1,0 +1,2 @@
+(define-keyset 'cp-module-admin
+  (read-keyset "module-admin-keyset"))
