+++ corpus/pact-examples/verified-accounts/accounts.repl
@@ -1,0 +1,7 @@
+(env-keys ["accountadmin" "user123" "user456"])
+(env-data { "accounts-admin-keyset": ["accountadmin"] })
+(begin-tx)
+(load "accounts.pact")
+(commit-tx)
+(verify 'accounts)
+; (typecheck 'accounts true)
