+++ lib/Cargo.toml
@@ -8,6 +8,7 @@ logos = "0.14.0"
 
 [dev-dependencies]
 criterion = "0.3"
+pactfmt-macros = { path = "../macros" }
 
 [[bench]]
 name = "benchmark"
