+++ lib/benches/benchmark.rs
@@ -1,14 +1,17 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use logos::Logos;
-use pactfmt_lib::Token;
+use pactfmt_lib::lexer::Token;
+use pactfmt_macros::benchmark_corpus;
 
-pub fn criterion_benchmark(c: &mut Criterion) {
+pub fn lexer_benchmark(c: &mut Criterion) {
+    let mut group = c.benchmark_group("lex");
+
     fn lex(input: &str) -> Vec<Result<Token, ()>> {
         Token::lexer(input).collect()
     }
 
-    c.bench_function("lex", |b| b.iter(|| lex(black_box("( )"))));
+    benchmark_corpus!(lex);
 }
 
-criterion_group!(benches, criterion_benchmark);
+criterion_group!(benches, lexer_benchmark);
 criterion_main!(benches);
