+++ macros/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "pactfmt-macros"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+quote = "1"
+proc-macro2 = "1"
+syn = "2"
+unicode-ident = "1"
